refactor: `ProductName` -> `Name` 로 변경 및 common.domain 패키지로 이동
