+++ src/main/java/kitchenpos/common/domain/Name.java
@@ -1,0 +1,49 @@
+package kitchenpos.common.domain;
+
+import java.util.Objects;
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+import kitchenpos.products.infra.PurgomalumClient;
+import kitchenpos.util.StringUtils;
+
+@Embeddable
+public class Name {
+
+    @Column(name = "name", nullable = false)
+    private String value;
+
+    protected Name() { }
+
+    public Name(String value, PurgomalumClient purgomalumClient) {
+        if (StringUtils.isBlank(value)) {
+            throw new IllegalArgumentException("이름은 빈 값이 될 수 없습니다.");
+        }
+
+        if (purgomalumClient.containsProfanity(value)) {
+            throw new IllegalArgumentException("이름에 비속어가 포함되어 있습니다.");
+        }
+
+        this.value = value;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Name that = (Name) o;
+        return value.equals(that.value);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
+    }
+
+    public String value() {
+        return value;
+    }
+}
